Fragmentation down to more manageable parts.
