+++ test/query/parameterized/test_project.py
@@ -2,6 +2,6 @@
 
 from __future__ import unicode_literals
 
-from marrow.mongo.param import F
+from marrow.mongo.param import P
 
-F  # Filter Document
+P  # Projection Document
